refator: remove PoseEstimator
